Code for 2024-05-13

+++ src/main/java/com/example/demo/models/UserPageModel.java
@@ -10,5 +10,5 @@ import java.util.List;
 public class UserPageModel {
     private List<UserModel> Users;
     private int totalPages;
-    private int totalElements;
+    private Long totalElements;
 }
